Added funmoney to account page, total monthly expenses for budget, renaming budgets, deleting budgets, and changing budgets amount.

+++ database/migrations/2020_03_31_193715_create_expenses_table.php
@@ -17,7 +17,7 @@ class CreateExpensesTable extends Migration
             $table->id();
             $table->integer('user');
             $table->integer('category');
-            $table->double('amount');
+            $table->integer('amount');
             $table->timestamps();
         });
     }
